requirements.txt
pyyaml
jinja2
reportlab
weasyprint
